Script is updated to cover all cases of INVALID_DATA for DYNAMIC image type, all parameters
